# prototipes/command-service/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API de escritura para el sistema de inventario basado en arquitectura CQRS + EDA",
        "title": "Command Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "carlosand_01@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/inventory/items": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo item en el inventario. El SKU debe ser único y la cantidad inicial debe ser \u003e= 0.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Create a new inventory item",
                "parameters": [
                    {
                        "description": "Item creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Item creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateItemResponse"
                        }
                    },
                    "400": {
                        "description": "Request inválido - campos requeridos faltantes o valores inválidos",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autorizado - token JWT inválido o faltante",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflicto - SKU duplicado",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor - error de persistencia o conexión a base de datos",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Servicio no disponible - error de conexión al event broker",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/inventory/items/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza un item existente en el inventario. Solo se pueden actualizar el nombre y la descripción.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Update an inventory item",
                "parameters": [
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "Item ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Item update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item actualizado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateItemResponse"
                        }
                    },
                    "400": {
                        "description": "Request inválido - ID inválido o campos requeridos faltantes",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autorizado - token JWT inválido o faltante",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Item no encontrado",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor - error de persistencia o conexión a base de datos",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Servicio no disponible - error de conexión al event broker",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina un item del inventario. Esta operación no se puede deshacer.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Delete an inventory item",
                "parameters": [
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "Item ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item eliminado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autorizado - token JWT inválido o faltante",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Item no encontrado",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor - error de persistencia o conexión a base de datos",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Servicio no disponible - error de conexión al event broker",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/inventory/items/{id}/adjust": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ajusta la cantidad de stock de un item. Valores positivos aumentan el stock, valores negativos lo disminuyen. No se puede ajustar a un valor negativo total.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Adjust stock quantity",
                "parameters": [
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "Item ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stock adjustment request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AdjustStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stock ajustado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/handlers.StockResponse"
                        }
                    },
                    "400": {
                        "description": "Request inválido - ID inválido, cantidad faltante o stock insuficiente",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autorizado - token JWT inválido o faltante",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Item no encontrado",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor - error de persistencia o conexión a base de datos",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Servicio no disponible - error de conexión al event broker",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateItemRequest": {
            "description": "Request to create a new inventory item",
            "type": "object",
            "required": [
                "name",
                "quantity",
                "sku"
            ],
            "properties": {
                "description": {
                    "description": "Product description (optional)\n@Example \"High-performance laptop with 16GB RAM and 512GB SSD\"\n@Example \"\"",
                    "type": "string",
                    "example": "High-performance laptop with 16GB RAM and 512GB SSD"
                },
                "name": {
                    "description": "Product name\n@Example \"Laptop Dell XPS 15\"\n@Example \"iPhone 15 Pro Max\"",
                    "type": "string",
                    "example": "Laptop Dell XPS 15"
                },
                "quantity": {
                    "description": "Initial stock quantity (must be \u003e= 0)\n@Example 100\n@Example 0\n@Example 500",
                    "type": "integer",
                    "minimum": 0,
                    "example": 100
                },
                "sku": {
                    "description": "SKU (Stock Keeping Unit) - unique identifier for the product\n@Example \"SKU-001\"\n@Example \"PROD-2024-ABC\"",
                    "type": "string",
                    "example": "SKU-001"
                }
            }
        },
        "handlers.CreateItemResponse": {
            "description": "Response after successfully creating an inventory item",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Creation timestamp (ISO 8601 format)",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "description": {
                    "description": "Product description",
                    "type": "string",
                    "example": "High-performance laptop with 16GB RAM and 512GB SSD"
                },
                "id": {
                    "description": "Unique item identifier (UUID)",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "description": "Product name",
                    "type": "string",
                    "example": "Laptop Dell XPS 15"
                },
                "quantity": {
                    "description": "Current stock quantity",
                    "type": "integer",
                    "example": 100
                },
                "sku": {
                    "description": "SKU (Stock Keeping Unit)",
                    "type": "string",
                    "example": "SKU-001"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}